fix: 점수 계산 방식 변경 & add dataStore di

+++ app/src/main/java/com/jbrunoo/digitink/domain/model/Result.kt
@@ -1,12 +1,12 @@
 package com.jbrunoo.digitink.domain.model
 
 data class Result(
-    val speedGame5: Int = 0,
-    val speedGame10: Int = 0,
-    val speedGame15: Int = 0,
-    val speedGame20: Int = 0,
+    val speedGame5: Long = 0,
+    val speedGame10: Long = 0,
+    val speedGame15: Long = 0,
+    val speedGame20: Long = 0,
 ) {
-    fun find(count: Int): Int {
+    fun find(count: Int): Long {
         return when(count) {
             5 -> speedGame5
             10 -> speedGame10
